Partially fix the game startup / tick # > 0 issue.
Previously, the game loop began ticking before the countdown finished
Now we check `if (this.countdownActive || !this.gameStarted)` in the tick fn

TODO: A lot more clean up of the code for game startup
TODO: We need to account for / fix the FE delay.
I'm stilling army count of 2 after the countdown finishes (we should see 1 to start).