Lots of deploy fixes... got initial deploy working!! TODO: revisit all of this
We had a lot of issues TS in the backend
Especially with the path aliases. Right now we are using ts-node,
instead of compiling and running the compiled file: `node server.js`.

Fixing DATABASE_URL (the real problem was a "/" in the db password)

Fix caddy config (had to tweak this multiple times...)
Subdomains would be easier / cleaner probably,
Due to the FE SPA needing to be the ultimate fallback at "index.html"