Refactor remaining files in common, moving some stuff to platform.
Right now we are leaning towards *not* having generic shared domain models
in platform, but I'm not sure how it will play out.
I wasn't sure how to refactor some of the last code in common/types, so
this is a good compromise for now that lets me finally delete common.